1.订单支付 使用余额支付 并未成功 但是跳转到了完成页面 2.进入app未提示有未付款项目 3.支付页面需要显示余额具体多少 4.行程中左上角返回键不要 5.声音寻车无图标 6.支付界面 显示可用优惠券xx张> 7.修改首页停车场车辆数据不刷新； 8.修改用车流程中起点终点的位置 自动缩放 9.修改所有车辆信息车辆图片； 10.修改优惠券有效期显示；

+++ Udive_az/app/src/main/java/cn/com/i_zj/udrive_az/model/UnFinishOrderResult.java
@@ -69,14 +69,14 @@ public class UnFinishOrderResult implements Serializable {
         private int startParkId;
         private int destinationParkId;
         private long startTime;
-        private int endTime;
+        private long endTime;
         private int shouldPayAmount;
         private int realPayAmount;
         private int ownerEarnAmount;
-        private int payTime;
+        private long payTime;
         private int payType;
         private int mileage;
-        private int durationTime;
+        private long durationTime;
         private int deductibleStatus;
         private int status;
         private int discountId;
@@ -140,11 +140,11 @@ public class UnFinishOrderResult implements Serializable {
             this.startTime = startTime;
         }
 
-        public int getEndTime() {
+        public long getEndTime() {
             return endTime;
         }
 
-        public void setEndTime(int endTime) {
+        public void setEndTime(long endTime) {
             this.endTime = endTime;
         }
 
@@ -172,7 +172,7 @@ public class UnFinishOrderResult implements Serializable {
             this.ownerEarnAmount = ownerEarnAmount;
         }
 
-        public int getPayTime() {
+        public long getPayTime() {
             return payTime;
         }
 
@@ -196,11 +196,11 @@ public class UnFinishOrderResult implements Serializable {
             this.mileage = mileage;
         }
 
-        public int getDurationTime() {
+        public long getDurationTime() {
             return durationTime;
         }
 
-        public void setDurationTime(int durationTime) {
+        public void setDurationTime(long durationTime) {
             this.durationTime = durationTime;
         }
